Migrate to Scala 3 (still TODO: DeriveArbitrary)
This migrates the production code to Scala 3.

Main change is that derivation of (circe) Json typeclasses requires more
boilerplate than before. If we only derive for the "root" message
types, derivation will fail, so this adds (derived) typeclass
instances for various "child" types.

This change does not fully migrate the tests yet. What's still pending
is the derivation of Arbitrary-instances to get the tests to run again.